Q5 adjustment + some other stuff
Adjusted the model settings in Q5 to produce a result with both Authors and Study as a random effect.

Doing this required setting the tolerance very high, as for some reason the setting to change the number of iterations wasnt working.

Q9 and 10 added to but unfinished

Some other unimportant stuff was removed